Refactored the Monster Uploader and updated the monsters to the most recent version
